Aggiunta l'azione edit sul teachers_controller e la relativa view. Per ora serve solamente per aggiungere o rimuove corsi di laurea ai docenti. Sto pensando se è meglio fare un'azione distinta per ogni tipo di modifica sul teacher
git-svn-id: http://subversion.assembla.com/svn/sigeol@182 080cfd25-d5d9-4c59-82ea-70d02a200d8b

+++ app/controllers/teachers_controller.rb
@@ -1,27 +1,63 @@
 class TeachersController < ApplicationController
   skip_before_filter :login_required, :only => [:index, :activate]
-  before_filter :manage_teachers_required, :only => [:new, :create, :administration]
+  before_filter :manage_teachers_required, :only => [:new, :create, :administration, :edit,
+                                                     :update]
+  before_filter :same_graduate_course_required, :only => [:edit, :update]
   def index
     @teachers = User.find_by_specified_type("Teacher")
   end
 
+  def show
+    @teacher = Teacher.find(params[:id])
+    respond_to do |format|
+      format.html # show.html.erb
+      format.xml  { render :xml => @teacher }
+    end
+  end
+
+  def edit
+    total_graduate_courses = @current_user.graduate_courses
+    @teacher = Teacher.find(params[:id])
+    teacher_graduate_courses = @teacher.user.graduate_courses
+    @manageable_graduate_courses = total_graduate_courses & teacher_graduate_courses
+    @selectable_graduate_courses = total_graduate_courses - teacher_graduate_courses
+  end
+
+  def update
+
+  end
+
   def new
     @teacher = User.new()
+    @graduate_courses = @current_user.graduate_courses
   end
 
   def create
-    user = User.new()
-    user.mail = params[:mail]
-    user.random = rand(120)
-    teacher = Teacher.new()
-    user.specified = teacher
-    if user.save
-      flash[:notice] = "Docente invitato con successo"
-      redirect_to new_teacher_url
+    user = User.find_by_mail(params[:mail])
+    if user == nil
+      user = User.new()
+      user.mail = params[:mail]
+      user.graduate_courses << GraduateCourse.find(params[:graduate_course_id])
+      user.random = rand(120)
+      teacher = Teacher.new()
+      user.specified = teacher
+      if user.save
+        flash[:notice] = "Docente invitato con successo"
+        redirect_to new_teacher_url
+      else
+        flash[:error] = user.errors
+        teacher.destroy
+        redirect_to new_teacher_url
+      end
     else
-      flash[:error] = user.errors.full_messages.to_s
-      teacher.destroy
-      redirect_to new_teacher_url
+      if user.own_by_teacher?
+        user.graduate_courses << GraduateCourse.find(params[:graduate_course_id])
+        flash[:notice] = "Docente invitato con successo"
+        redirect_to new_teacher_url
+      else
+        flash[:errors] = "La mail inserita è già presente e non corrisponde ad un docente"
+        redirect_to new_teacher_url
+      end
     end
   end
 
@@ -60,5 +96,18 @@ class TeachersController < ApplicationController
   def administration
     ids = @current_user.graduate_course_ids
     @graduate_courses = GraduateCourse.find(ids, :include => [:users] , :conditions => "users.specified_type = 'Teacher'")
+  end
+
+  private
+
+  def same_graduate_course_required
+    teacher = Teacher.find(params[:id])
+    teacher_ids = teacher.user.graduate_course_ids
+    current_user_ids = @current_user.graduate_course_ids
+    common_ids = current_user_ids & teacher_ids
+    if common_ids == []
+      flash[:error] = "Questo docente non appartiane a nessun tuo corso di laurea"
+      redirect_to timetables_url
+    end
   end
 end
